Propagate compose State when loading an icon
This change fixes the case when we need to change an icon of the same button. The new drawable produced by the loadDrawable haven't been propogated to loadIcon and eventually Icon. We also can't use loadDrawable function in loadIcon because the former returns a cached drawable before loading the new one, which then gets assigned the resId of the new drawable that is being loaded and later messes up with Icon#equals function.

Flag: com.android.systemui.new_screen_record_toolbar
Bug: 28686600
Test: atest DrawablesViewModelTest
Change-Id: I41d4232b87c59b9f366cb7bfa6d56c9a4b9dc113