Update test TableLogBuffers to use LogProxyDefault.
Unit tests will automatically echo android.util.Log statements to the
output, and we want TableLogBuffer content to be part of that output as
well. FakeLogProxy doesn't do that echo-ing, so we actually want the
real LogProxyDefault in tests so the buffer is logged to logcat and
available in test runs.

Bug: EXEMPT test fix
Flag: TEST_ONLY
Test: atest BatteryRepositoryTest -> verify TableLogBuffer content is in
output in studio
Test: atest TableLogBufferFactoryImplTest still passes

Change-Id: Icb06bf4d7961e0dcf1c1f41444c95343852da8e1