Dismiss PiP without bespoke animation when it is not visible.
Previously the default behavior was to stay in PiP when a transition
brought the PiP'ed activity to the front, if PiP was not visible
before the transition (i.e. when launched from the lockscreen).

However we want notification taps for the PiP'ed app to always go
fullscreen, even from the lockscreen. This change expands PiP
instead of keeping it open when coming from the lockscreen, and
crucially does not mark the PiP transition handler as the default for
the transition, so that the keyguard dismiss transition can handle
it instead.

Bug: 373314569
Flag: com.android.wm.shell.dismiss_pip_from_lockscreen
Test: atest PipSchedulerTest + manual (see videos in the bug)
Change-Id: I01dd7f35bce1b989f2d2df8c65ab3f0f890379d5