[WindowDecor] Add null check for AppHeaderViewHolder
in onMaximizeButtonHoverExit. In case we get an additional motion event
after the window decoration is closed/destroyed, we might get a
onMaximizeButtonHoverExit call after mWindowDecorViewHolder is already
closed and set to null, which will lead to an NPE.

This CL adds a null check to prevent the crash if this happens. This is
not needed in WindowDecor2 because the view holder is closed but never
set to null there.

Fixes: 435664004
Test: atest DesktopModeWindowDecorationTests
Flag: EXEMPT bug fix
Change-Id: I1cd26cd14ec7c37074d66d891f2c9b828da60c7f