Remove cooldown for sessionless foreground vol key routing
Previous to this change, vol keys would be routed to the foreground
app if it was the last app to play audio, even if that was hours ago
(and therefore far outside the user's expectation/context). This change
reduces this infinite timeout to zero, so as soon as the foreground
stops playing audio the keys will be routed back to the remote/casting
session. This is aligned with other volume key routing decisions in
AudioService (zero cooldown).

Bug: 432003816
Test: Manual testing & check MediaSessionTest and MediaActivityTest pass with flag enabled
Flag: com.android.media.mediasession.flags.sessionless_vol_key_zero_cooldown
Change-Id: I3b952997d4b77593df598474f215e133199a49d2