Replace @pending with @hide pending
Supporting both `@pending` and `@hide` makes Metalava slower and more
complex and `@pending` does not provide any real benefit over just
using `@hide`, particularly when it can be followed with text
explaining why it is `@hide`.

This change replaces the two occurences of `@pending` that affect the
API surface with `@hide pending`. It also moves the `@hide` to the
block tag section after the main description.

Flag: EXEMPT minor refactoring to simplify Metalava
Bug: 429965593
Test: m checkapi
Change-Id: I966f5c0c579a8d350a117db60bb0a207970c271f