[aapt2] Fix the autoshifting sdk version <36
The old code used to update the parsed config's sdk version to
be at least Baklava (36) the same way we do it for other
components - assuming we need to make the whole configuration
target the first SDK release with the feature support.

This caused typos (e.g. 30.1) to be 'autocorrected' to 36.1
instead of letting the developer know about the typo. Given that
any version with minor set and sdkVersion <36 is just not valid,
it's safe to return an error instead at all times.

Bug: 438788698
Flag: EXEMPT host tools don't support flags
Test: atest aapt2_tests libandroidfw_tests
Change-Id: I890e3b1f2047f45030b865ed036221ff6d2a3e83