Desktop screenshots: Refactor capture method after user click
When the user clicks the capture button either for screenshots or screen
recording, it consolidates to a single method which then determines
which action to take based on the current state.

Additionally, after ag/34814866 landed, this fixes some comments so that
they no longer refer to the obsolete activity.

Also fixes the casing convention for various methods and members to
adhere to the Kotlin style guide.

Bug: 435225255
Test: atest PreCaptureViewModelTest
Flag: com.android.systemui.large_screen_screencapture
Change-Id: Ibe79ed0eddbb728cb29c456ea599e0df0196d5aa