Remove the Job category APIs.
The job category APIs were introduced to allow clients to
specify a category for their jobs, with the intention of
influencing scheduling and resource allocation. However,
the plan for this feature has changed, and we are not
moving forward with it at this time.

NO_EXPORTED_FLAG_DELETION_CHECK=This API is never released
and no client is relying on this.

Bug: 419047126
Bug: 439641599
Test: n/a
Flag: EXEMPT removing android.app.job.job_category_apis
Change-Id: I7833c6de694c5153bc2735203f2c07ead282b00e