Remove addWindowToken from WindowToken constructor
Currently when a new WindowToken is created, it will be added to the
given DisplayContent (if any) from its constructor. This creates some
issues, as adding it to the DisplayContent will immediately create its
surface, which cannot be changed anymore. If some class extends the
WindowToken, adds custom fields, and tries to use those fields in the
title of the WindowToken, they will show default values. This is because
the title is only initialized once, with the surface, during the super
constructor, so any values set afterwards are ignored.

This extracts the addWindowToken call out of the constructor, and into
the caller sites, making the intent more explicit, as well as keeping
the constructor pure.

Flag: EXEMPT BUGFIX
Bug: 281029564
Test: atest DisplayAreaPolicyBuilderTest
  DualDisplayAreaGroupPolicyTest
  TransitionTests
  WallpaperControllerTests
  WindowContextListenerControllerTests
  WindowTokenTests
Change-Id: I38b95cfc707c4da785c93e350371c419c45512f0