Refactor isTaskMaximized to enforce caller to validate DisplayLayout
isTaskMaximized implicitly did a null check (which could lead to crash)
when taskInfo.displayId is no longer valid. As this is just a boolean
function, returning null when display is missing is not a very clear. So
it's better to enforce from the caller side to pass DisplayLayout so it
will do a null check before passing it to isTaskMaximized

Also removed the redundant overloading

Bug: 426465402
Test: m & presubmit
Flag: EXEMPT bugfix
Change-Id: I43fbc75731a6f5fae878bbeda5fa1271ced8635e