Allow moving shade window from workspace swipe even when SceneContainer is disabled.
The logic to move the shade window is now only gated on
ShadeWindowGoesAround being enabled. This allows the shade window
to be moved with a workspace swipe even when the scene window container flag is disabled.

It's unclear why we gated this to also SceneContainer at first: it was probably for some flickers, but it seems they don't happen anymore.

Bug: 378685769
Test: scroll from workspace on external display while scene container flag is disabled
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I9faf990d15b1b07ed7bfa83d1d891a6e69405f1e