No footer message when showing empty shade
In dual shade, we show the footer even if the empty shade is visible.
Since the empty shade shows the "Unlock to see older notifications"
already, we don't need to show it in the footer in this case.

We don't need to check for dual shade especially, since the footer just
won't be visible at all when there are no notifications and dual shade
is off, so it doesn't matter.

Fix: 439537988
Test: FooterViewModelTest
Test: tested manually with multiple states (dual shade on/off), new
empty shade design on/off, with and without notifications
Flag: com.android.systemui.scene_container

Change-Id: Ib55b33e2b84291c28b1ae2e6201c48eb2b34d6aa