Add support for USB4/Thunderbolt tunnel management
This implements the base structure for USB4/Thunderbolt PCI tunnel
management. The goal is to add a new manager to USB service for enabling
or disabling PCI tunnels in the system and notifying the native
implementation of the screen lock state (by registering for keyguard
callbacks).

Bug: 369430205
Flag: com.android.server.usb.flags.enable_usb4
Test: Manual tests with flag disabled and enabled

Change-Id: I1cee70624e5a9c65ba8515ac2b79452936fa8dfe