Ensure expanding a bubble bar chat bubble hides the notification
When a bubble is added to the stack, it naturally becomes the
selected one. When expanding bubble bar from launcher, we'd skip
marking the bubble as accessed since the selection wasn't changed.

We don't want to skip this! This CL removes the check causing the skip.

Additionally:
- renames this method to indicate that it sets the expanded state
- adds tests for it
- adds to some existing tests to also check the showInShade state too

Flag: com.android.wm.shell.enable_bubble_bar
Test: atest BubbleDataTest
Bug: 428026020
Change-Id: Ib88bdb2ea5a5c541f05de58d40f7dcc7388738c9