Shortcut double-relayout under specific conditions
Under specific conditions, WM and VRI can coordinate to
avoid making a second sync-relayout when they "know" that
the second relayout will produce the same frames.

The specific conditions, in this case, are system windows
which fill the screen. In this situation, relayout ignores
configuration-dependent parameters, so even if there is a
pending configuration, we can be pretty sure that a
subsequent relayout call will produce the same frames. In
this way, we can allow the client to draw a frame early.

Bug: 439963876
Bug: 385976595
Test: Existing tests
Flag: com.android.window.flags.always_seq_id_layout
Change-Id: I92049d1582570a340f533e630ee41f500bd1c41a