Make PooledWindowDecorViewHostSupplier per-display.
Since we create PooledWindowDecorViewHostSupplier right at init, we use
the default context provided by the system - this is also used to create
new ViewHost instances. Instead of doing so, this CL allows creation of
Pools on Display add/remove callbacks, and keep a map of them to keep
the instances at a per-display level. This prevents ViewHosts that
should actually be created with a displayId != 0 to be created with
displayId = 0, which can lead to problems.

Bug: 440247561
Test: Manual. Wait for main display to turn off then interact with app
header - responds to hover events, and relayouts correctly when enter
fullscreen (become app handle).
Test: atest PooledWindowDecorViewHostSupplierTest
Flag:com.android.window.flags.enable_per_display_window_decor_view_host_pool

Change-Id: Icba5f48c21e6e94e768c0d082a3421c54935eaf1