Change namespace for use_default_visibility_for_sensitive_inputs
Changes the namespace of the flag to "systemui" as the changes guarded
by the flag are more UI related. As per go/flag-namespace-change, the
namespace of the flag cannot be changed unless the flag is renamed. So,
renaming the flag to enable_default_visibility_for_sensitive_inputs.

Bug: 339270220
Test: Builds correctly
Flag: com.android.internal.widget.flags.enable_default_visibility_for_sensitive_inputs
Change-Id: I6885591db39954514be078d4d1c1880fac753e4e