Keep transparent fullscreen task when that's the one being launched
Before this CL, when running a transition that launches a new task in a
desk, we would close the current top transparent fullscreen task in that
desk. With this CL, we don't close the top transparent fullscreen task
if that's the task that is being opened / moved to front in that desk.

Bug: 440096166
Flag: EXEMPT small bug fix
Test: DesktopTasksControllerTest, and manual
Change-Id: Icaaa63eacf740ef2760a0f31de16b900f91e60b5