Fall back shade window to default display when mirroring is enabled
When content mirroring is active, the shade should consistently appear on the default display, irrespective of any active ShadeDisplayPolicy.

This fixes the issue of shade being moved while mirroring, because we were using the display Id from the MotionEvent to decide where it was. On external displays with touchscreen, expanding the shade on the external display was causing the shade window to be reparented there

Bug: 441689533
Test: ShadeDisplaysRepositoryTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Ic0cafa532220c95c2fc3e010238efd91ed472520