ENR: do not skip resetting content alpha when alpha is 0
setHideSensitive alpha optimization logic skips resetting alpha when
showingPublic is not changed. However, showingLayout's content alpha is 0 in some edge
cases and it is reset correctly. This causes an experience which
Notification content looks completely blank.

This CL checks the showing layout content alpha and if it is 0, reset
content alphas in order to make sure content is shown correctly.

Bug: 441279545
Test: Presubmit
Test: manual (unlock the device with face and see hide sensitive flow
works as is)
Flag: EXEMPT trivial change. It's a check on top of alpha optimization
to make sure the view has correct alpha.

Change-Id: I9f0413e39c6bf1c9a834fa4d857f15a231963455