Fixing onDataSetChanged not getting called when the service factory was being created
The service was caching the serviceFactory and not the Wrapper around
the factory. So any internal state including mCreated was not cached.
Also mCreated was initialized as false for the first time and true for
cache-lookups, where as it should have been the other way round. This
caused the first call to the service not return correct data, if the
app was not doing any caching of its own.

Bug: 439638559
Flag: android.appwidget.flags.remote_adapter_conversion
Test: Verified various legacy list based widgets
Change-Id: Ib1b0d1530afe9fb3be35632363bb521d44fef595