Add PIN digits if quickly typed during reset
The reset animation introduces some delay when processing
PIN touches that can lead to a visual inconsistency: The touch
was recorded but not shown to the user. This could be confusing
to the user when they have correctly typed their PIN but it was
rejected. Make sure to visually show all taps after the reset
has completed.

Fixes: 441008042
Test: Manual - rapidly enter PIN digits with auto-entry on
Flag: EXEMPT bugfix
Change-Id: I4c0cfe4059121a3a854443076d7643963ecbf788