Fix IME back callback not being invoked for hardware back key
When an application has predictive back animations disabled
(enableOnBackInvokedCallback=false), but the active IME has them
enabled, the IME's registered back callback was not being invoked for a
hardware back key event.

Instead of invoking the callback, the system would fall back to the
default behavior of hiding the IME. While this is often the desired
outcome, it prevents IMEs from implementing custom back handling logic.

This change ensures that if the top back callback belongs to the IME, it
is invoked regardless of the application's predictive back setting.

Bug: 442771291
Test: KeyboardVisibilityControlTest
Flag: EXEMPT bugfix
Change-Id: I5dffe289ca73c3fd038284f85dc5f9d6762251c6