a11y: Update getMagnificationModeLocked in userState
Before this change, for a specific display, if there is no cached mode
on the display, this function will return full screen mode.

After this change, if there is no cached mode on the display, and the
default display has a cached mode, then it will return the cached mode
on the default display.

Bug: b/442821756
Flag: EXEMPT update existing method
Test: atest AccessibilityUserStateTest
Change-Id: I2b842d7e9934941c993ccb0a506b026949df06e7