BatchedInputEventReceiverTest - mock Choreographer
Updates BatchedInputEventReceiverTest to mock the choreographer used by
BatchedInputEventReceiverTest. While mocking a final class did not work
with the test implemented in kotlin, it seems to work if the test gets
converted to java. This allows us to remove "Scheduler" interface used
to fake interactions between BatchedInputEventReceiver and
Choreographer.

Bug: 444249275
Test: atest BatchedInputEventReceiver
Flag: EXEMPT TEST_ONLY

Change-Id: Iffb5e2e80f8a4132b6a8be48f79278025468a7ba