Increase scan + connection timeout for suggested device to 30 seconds.
The previous timeout was 10 seconds but supposed to be the timeout for
the scan only. However it gets reset not when the scan finishes, but
when the connection to a new device is established.

This is a quick workaround, while the fix that separates scan and
connection timeouts is behind the flag
`use_suggested_device_connection_manager`.

This change also fixes failing unit tests in LocalMediaManagerTest.

Bug: 443107229
Test: atest LocalMediaManagerTest
Flag: EXEMPT BUGFIX
Change-Id: I7a8eaf637931c3f5974e5c072d1a522abc3ff8f6